fix build failure
