+++ src/include/test/unit_api_feature_vector.cc
@@ -31,6 +31,7 @@
 
 #include "api/feature_vector.h"
 #include "catch2/catch_all.hpp"
+#include "test/randomize.h"
 #include "test/test_utils.h"
 #include "utils/utils.h"
